refactor(gui): extract CommandPalette utilities and search hook
- Extract parseVersion/compareVersions to lib/version.ts
- Add comprehensive tests for version parsing and comparison
- Create useCommandSearch hook to encapsulate:
  - SearchItem union type (favorites + GVKs)
  - Unified fuzzy search across both types
  - Result separation and GVK sorting logic
- Reduce CommandPalette.tsx from 522 to 401 lines (23% reduction)
